backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Justin Hsie <justinhsie1001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
httpx = "^0.22.0"
gunicorn = "^20.1.0"
asyncpg = "^0.25.0"
databases = "^0.5.5"
psycopg2-binary = "^2.9.3"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
